requirements.txt
psutil==7.0.0
python-dotenv==1.1.1
requests==2.32.5
